requirement.txt

Pillow==9.4.0
python-pptx==0.6.21

